renames application to Lambda Social, previously it was just Client which is stupid
