+++ client/android/app/BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.client",
+    package = "com.lambdasocial",
 )
 
 android_resource(
     name = "res",
-    package = "com.client",
+    package = "com.lambdasocial",
     res = "src/main/res",
 )
