+++ client/android/app/build.gradle
@@ -102,7 +102,7 @@ android {
     }
 
     defaultConfig {
-        applicationId "com.client"
+        applicationId "com.lambdasocial"
         minSdkVersion rootProject.ext.minSdkVersion
         targetSdkVersion rootProject.ext.targetSdkVersion
         versionCode 1
