+++ client/android/settings.gradle
@@ -1,4 +1,4 @@
-rootProject.name = 'client'
+rootProject.name = 'LambdaSocial'
 include ':react-native-auth0'
 project(':react-native-auth0').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-auth0/android')
 include ':react-native-screens'
