+++ client/app.json
@@ -1,4 +1,4 @@
 {
-  "name": "client",
+  "name": "LambdaSocial",
   "displayName": "Lambda Social"
 }
