+++ client/ios/LambdaSocial-tvOS/Info.plist
