+++ client/ios/LambdaSocial-tvOSTests/Info.plist
