+++ client/ios/LambdaSocial.xcodeproj/xcshareddata/xcschemes/LambdaSocial-tvOS.xcscheme
@@ -29,9 +29,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-               BuildableName = "client-tvOS.app"
-               BlueprintName = "client-tvOS"
-               ReferencedContainer = "container:client.xcodeproj">
+               BuildableName = "LambdaSocial-tvOS.app"
+               BlueprintName = "LambdaSocial-tvOS"
+               ReferencedContainer = "container:LambdaSocial.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
          <BuildActionEntry
@@ -43,9 +43,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "2D02E48F1E0B4A5D006451C7"
-               BuildableName = "client-tvOSTests.xctest"
-               BlueprintName = "client-tvOSTests"
-               ReferencedContainer = "container:client.xcodeproj">
+               BuildableName = "LambdaSocial-tvOSTests.xctest"
+               BlueprintName = "LambdaSocial-tvOSTests"
+               ReferencedContainer = "container:LambdaSocial.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -61,9 +61,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "2D02E48F1E0B4A5D006451C7"
-               BuildableName = "client-tvOSTests.xctest"
-               BlueprintName = "client-tvOSTests"
-               ReferencedContainer = "container:client.xcodeproj">
+               BuildableName = "LambdaSocial-tvOSTests.xctest"
+               BlueprintName = "LambdaSocial-tvOSTests"
+               ReferencedContainer = "container:LambdaSocial.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -71,9 +71,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "client-tvOS.app"
-            BlueprintName = "client-tvOS"
-            ReferencedContainer = "container:client.xcodeproj">
+            BuildableName = "LambdaSocial-tvOS.app"
+            BlueprintName = "LambdaSocial-tvOS"
+            ReferencedContainer = "container:LambdaSocial.xcodeproj">
          </BuildableReference>
       </MacroExpansion>
       <AdditionalOptions>
@@ -94,9 +94,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "client-tvOS.app"
-            BlueprintName = "client-tvOS"
-            ReferencedContainer = "container:client.xcodeproj">
+            BuildableName = "LambdaSocial-tvOS.app"
+            BlueprintName = "LambdaSocial-tvOS"
+            ReferencedContainer = "container:LambdaSocial.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
       <AdditionalOptions>
@@ -113,9 +113,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "client-tvOS.app"
-            BlueprintName = "client-tvOS"
-            ReferencedContainer = "container:client.xcodeproj">
+            BuildableName = "LambdaSocial-tvOS.app"
+            BlueprintName = "LambdaSocial-tvOS"
+            ReferencedContainer = "container:LambdaSocial.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
