+++ client/ios/LambdaSocial/AppDelegate.h
