+++ client/ios/LambdaSocial/Images.xcassets/AppIcon.appiconset/Contents.json
