+++ client/ios/LambdaSocial/Images.xcassets/Contents.json
