+++ client/ios/LambdaSocialTests/Info.plist
