+++ client/ios/LambdaSocialTests/LambdaSocialTests.m
@@ -14,11 +14,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React Native!"
 
-@interface clientTests : XCTestCase
+@interface LambdaSocialTests : XCTestCase
 
 @end
 
-@implementation clientTests
+@implementation LambdaSocialTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL(^)(UIView *view))test
 {
